requirements.txt
Flask
gunicorn
eventlet
python-socketio
psycopg2-binary
sqlalchemy
fpdf